requirements.txt
asgiref==3.8.1
cffi==1.17.1
chardet==5.2.0
cryptography==44.0.0
cssselect2==0.7.0
Django==5.1.4
django-extensions==3.2.3
django-taggit==6.1.0
easy-thumbnails==2.10
lxml==5.3.0
Markdown==3.4.1
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
pyOpenSSL==24.3.0
reportlab==4.2.5
sqlparse==0.5.2
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.12.2
webencodings==0.5.1
